Create index, first recipe and an "img" directory
Here i created the main index page where all recipes will be linked.
A directory for the images I will use.
And a ".html" page of the first recipe, "Italian Chicken Marinade".
With the name of the recipe, a picture of the finished recipe, a basic
description, list of ingredients and a list of steps to follow.